Gpt/requirements.txt
discord.py==1.7.3
Flask==2.1.4
